+++ test/unit/card-number.js
@@ -1,16 +1,19 @@
+'use strict';
+
 var expect = require('chai').expect;
 var cardNumber = require('../../src/card-number');
 
 describe('number validates', function () {
-
   describe('partial validation sequences', function () {
     table([
       ['',
         {card: null, isPotentiallyValid: true, isValid: false}],
       ['6',
         {card: null, isPotentiallyValid: true, isValid: false}],
+      ['60',
+        {card: null, isPotentiallyValid: true, isValid: false}],
       ['601',
-        {card: 'discover', isPotentiallyValid: true, isValid: false}],
+        {card: null, isPotentiallyValid: true, isValid: false}],
       ['6011',
         {card: 'discover', isPotentiallyValid: true, isValid: false}],
       ['4',
@@ -24,7 +27,7 @@ describe('number validates', function () {
       ['x',
         {card: null, isPotentiallyValid: false, isValid: false}],
       ['123',
-        {card: null, isPotentiallyValid: false, isValid: false}],
+        {card: null, isPotentiallyValid: false, isValid: false}]
     ]);
   });
 
@@ -52,8 +55,8 @@ describe('number validates', function () {
         {card: null, isPotentiallyValid: false, isValid: false}],
       ['41111111111111111111', // Too long
         {card: 'visa', isPotentiallyValid: false, isValid: false}],
-      ['4111111111111112', // right lenght, not Luhn
-        {card: 'visa', isPotentiallyValid: false, isValid: false}],
+      ['4111111111111112', // right length, not Luhn, potentialls valid because visas can be 19 digits
+        {card: 'visa', isPotentiallyValid: true, isValid: false}]
     ]);
   });
 
@@ -64,34 +67,41 @@ describe('number validates', function () {
       ['4111 1111 1111 1111',
         {card: 'visa', isPotentiallyValid: true, isValid: true}],
       ['601 1 1 1  1 1 1 1   1 1 1 1 1 7',
-        {card: 'discover', isPotentiallyValid: true, isValid: true}],
+        {card: 'discover', isPotentiallyValid: true, isValid: true}]
     ]);
   });
 
   describe('Discover', function () {
     table([
-      ['60',
-        {card: 'discover', isPotentiallyValid: true, isValid: false}],
       ['6011111',
         {card: 'discover', isPotentiallyValid: true, isValid: false}],
       ['6011111111111117',
-        {card: 'discover', isPotentiallyValid: true, isValid: true}],
+        {card: 'discover', isPotentiallyValid: true, isValid: true}]
     ]);
   });
 
   describe('MasterCard', function () {
     table([
+      ['2',
+        {card: null, isPotentiallyValid: true, isValid: false}],
+      ['27',
+        {card: 'mastercard', isPotentiallyValid: true, isValid: false}],
+      ['272',
+        {card: 'mastercard', isPotentiallyValid: true, isValid: false}],
+      ['2720',
+        {card: 'mastercard', isPotentiallyValid: true, isValid: false}],
       ['55555555',
-        {card: 'master-card', isPotentiallyValid: true, isValid: false}],
+        {card: 'mastercard', isPotentiallyValid: true, isValid: false}],
       ['5555555555554444',
-        {card: 'master-card', isPotentiallyValid: true, isValid: true}],
+        {card: 'mastercard', isPotentiallyValid: true, isValid: true}],
       ['5555555555554446',
-        {card: 'master-card', isPotentiallyValid: false, isValid: false}],
+        {card: 'mastercard', isPotentiallyValid: false, isValid: false}]
     ]);
   });
 
   describe('Maestro', function () {
     table([
+      ['602011', {card: 'maestro', isPotentiallyValid: true, isValid: false}],
       ['500000000000', {card: 'maestro', isPotentiallyValid: true, isValid: false}],
       ['500000000000061', {card: 'maestro', isPotentiallyValid: true, isValid: false}],
       ['5000000000000611', {card: 'maestro', isPotentiallyValid: true, isValid: true}],
@@ -99,7 +109,7 @@ describe('number validates', function () {
       ['500000000000000005', {card: 'maestro', isPotentiallyValid: true, isValid: false}],
       ['5000000000000000005', {card: 'maestro', isPotentiallyValid: true, isValid: true}],
       ['5000000000000000001', {card: 'maestro', isPotentiallyValid: false, isValid: false}],
-      ['50000000000000000009', {card: 'maestro', isPotentiallyValid: false, isValid: false}],
+      ['50000000000000000009', {card: 'maestro', isPotentiallyValid: false, isValid: false}]
     ]);
   });
 
@@ -108,7 +118,7 @@ describe('number validates', function () {
       ['3782',
         {card: 'american-express', isPotentiallyValid: true, isValid: false}],
       ['378282246310005',
-        {card: 'american-express', isPotentiallyValid: true, isValid: true}],
+        {card: 'american-express', isPotentiallyValid: true, isValid: true}]
     ]);
   });
 
@@ -116,12 +126,12 @@ describe('number validates', function () {
     table([
       ['1',
         {card: 'jcb', isPotentiallyValid: true, isValid: false}],
-      ['2',
+      ['21',
         {card: 'jcb', isPotentiallyValid: true, isValid: false}],
       ['3530111',
         {card: 'jcb', isPotentiallyValid: true, isValid: false}],
       ['3530111333300000',
-        {card: 'jcb', isPotentiallyValid: true, isValid: true}],
+        {card: 'jcb', isPotentiallyValid: true, isValid: true}]
     ]);
   });
 
@@ -144,10 +154,9 @@ describe('number validates', function () {
       [true,
         {card: null, isPotentiallyValid: false, isValid: false}],
       [false,
-        {card: null, isPotentiallyValid: false, isValid: false}],
+        {card: null, isPotentiallyValid: false, isValid: false}]
     ]);
   });
-
 });
 
 function table(tests) {