+++ test/unit/cardholder-name.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var expect = require('chai').expect;
 var postalCode = require('../../src/cardholder-name');
 
@@ -6,7 +8,7 @@ describe('postalCode', function () {
     'returns false for non-string types': [
       [0, {isValid: false}],
       [1, {isValid: false}],
-      [-1,{isValid: false}],
+      [-1, {isValid: false}],
       [null, {isValid: false}],
       [[], {isValid: false}],
       [{}, {isValid: false}]
@@ -15,18 +17,18 @@ describe('postalCode', function () {
     'accepts strings of any length': [
       ['Herbert', {isValid: true}],
       ['Michael Knight', {isValid: true}],
-      ['  Garth    !!!  ', {isValid: true}],
-
+      ['  Garth    !!!  ', {isValid: true}]
     ],
 
     'returns false on empty strings': [
       ['', {isValid: false}],
-      [' ', {isValid: false}],
+      [' ', {isValid: false}]
     ]
   };
 
   Object.keys(describes).forEach(function (key) {
     var tests = describes[key];
+
     describe(key, function () {
       tests.forEach(function (test) {
         var arg = test[0];
@@ -34,9 +36,8 @@ describe('postalCode', function () {
 
         it('returns ' + JSON.stringify(output) + ' for "' + arg + '"', function () {
           expect(postalCode(arg)).to.deep.equal(output);
-        })
+        });
       });
     });
   });
-
 });