+++ test/unit/expiration-month.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var expect = require('chai').expect;
 var expirationMonth = require('../../src/expiration-month');
 
@@ -14,7 +16,7 @@ describe('expirationMonth', function () {
       [[], FALSE_VALIDATION],
       [{}, FALSE_VALIDATION],
       [null, FALSE_VALIDATION],
-      [undefined, FALSE_VALIDATION],
+      [undefined, FALSE_VALIDATION], // eslint-disable-line no-undefined
       [Infinity, FALSE_VALIDATION],
       [0 / 0, FALSE_VALIDATION],
       [0, FALSE_VALIDATION],
@@ -91,20 +93,24 @@ describe('expirationMonth', function () {
       ['-6', FALSE_VALIDATION],
       ['20', FALSE_VALIDATION],
       ['-1', FALSE_VALIDATION],
-      ['13', FALSE_VALIDATION],
-      [
-        previousMonth.toString(),
-        {
-          isValid: currentMonth !== 1,
-          isPotentiallyValid: true,
-          isValidForThisYear: false
-        }
-      ]
+      ['13', FALSE_VALIDATION]
     ]
   };
 
+  if (currentMonth !== 1) {
+    describes['invalid month'].push([
+      previousMonth.toString(),
+      {
+        isValid: currentMonth !== 1,
+        isPotentiallyValid: true,
+        isValidForThisYear: false
+      }
+    ]);
+  }
+
   Object.keys(describes).forEach(function (key) {
     var tests = describes[key];
+
     describe(key, function () {
       tests.forEach(function (test) {
         var arg = test[0];
@@ -112,7 +118,7 @@ describe('expirationMonth', function () {
 
         it('returns ' + JSON.stringify(output) + ' for "' + arg + '"', function () {
           expect(expirationMonth(arg)).to.deep.equal(output);
-        })
+        });
       });
     });
   });